feat: 최대 10개 검색 기능 추가
