+++ src/features/search/service/index.ts
@@ -1,3 +1,4 @@
 export { default as searchApi } from './handler';
 export { default as searchQueryOption } from './queryOption';
+export { default as searchLocalStorage } from './searchLocalStorage';
 export * from './types';
